Add support for deleting a product in list

+++ src/app/components/admin/admin-products/admin-products.component.ts
@@ -18,6 +18,7 @@ export class AdminProductsComponent implements OnInit {
   }
 
   delete(productId: string) {
+    if (!confirm('Are you sure you want to delete this item ? ')) { return; }
     this.productService.delete(productId);
   }
